Finish renaming work
